+++ src/tablecloth/api/missing.clj
@@ -52,4 +52,3 @@
      (process-group-data ds #(fill-range-replace % colname max-span missing-strategy missing-value))
      (dm/fill-range-replace ds colname max-span missing-strategy missing-value))))
 
-